+++ pkg/kubernetes/kubernetes.go
@@ -2,7 +2,6 @@ package kubernetes
 
 import (
 	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
 )
 
@@ -14,21 +13,23 @@ func (c *Client) GetClient() *kubernetes.Clientset {
 	return c.client
 }
 
-func NewClient(masterURL string, kubeconfig string) (*Client, error) {
+func NewClient(kubecontext string, kubeconfig string) (*Client, error) {
 
-	config, err := rest.InClusterConfig()
-	if err != nil {
-		kubeconfig :=
-			clientcmd.NewDefaultClientConfigLoadingRules().GetDefaultFilename()
-		config, err = clientcmd.BuildConfigFromFlags(masterURL, kubeconfig)
-		if err != nil {
-			return nil, err
-		}
-	}
-	clientSet, err := kubernetes.NewForConfig(config)
+	config := clientcmd.NewNonInteractiveDeferredLoadingClientConfig(
+		&clientcmd.ClientConfigLoadingRules{ExplicitPath: kubeconfig},
+		&clientcmd.ConfigOverrides{
+			CurrentContext: kubecontext,
+		})
+	// create the clientset
+	c, err := config.ClientConfig()
 	if err != nil {
 		return nil, err
 	}
+	clientSet, err := kubernetes.NewForConfig(c)
+	if err != nil {
+		return nil, err
+	}
+
 	return &Client{
 		client: clientSet,
 	}, nil